+++ src/index.js
@@ -1,12 +1,16 @@
 'use strict';
 
 const { score } = require('./scorer');
-const { tokenize, extractIntent, clearCache } = require('./tokenizer');
+const { tokenize, extractIntent, clearCache, resetSynonyms, loadSynonyms } = require('./tokenizer');
 const { loadSkills, parseSkillFrontmatter, discoverSkills, buildSkillIndex, detectProjectContext } = require('./scanner');
 const { setupAgentsMd } = require('./setup');
+const { computeSemanticScore, computeEmbedding, cosineSimilarity, computeSkillHash } = require('./semantic-scorer');
+const { createReranker } = require('./reranker');
 
 module.exports = {
   score, tokenize, loadSkills, extractIntent, parseSkillFrontmatter,
   discoverSkills, buildSkillIndex, detectProjectContext, setupAgentsMd,
-  clearCache
+  clearCache, resetSynonyms, loadSynonyms,
+  computeSemanticScore, computeEmbedding, cosineSimilarity, computeSkillHash,
+  createReranker
 };